+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- BREAKING: `Terminal:send` takes options as a table instead of ordered parameters.
+
 ## [0.0.2] - 2025-07-19
 
 - `Terminal:send` to support selection types in the `input` parameter: `single_line`, `visual_lines` or `visual_selection`.